# docs/swagger.yaml
basePath: /swagger/datahub-service
definitions:
  controllers.APIResponse:
    properties:
      data: {}
      msg:
        example: 操作成功
        type: string
      status:
        example: 0
        type: integer
    type: object
...
  models.JSONB:
    additionalProperties: true
    type: object
...
  models.TableField:
    properties:
      check_constraint:
        type: string
      data_type:
        type: string
      default_value:
        type: string
      description:
        type: string
      is_nullable:
        type: boolean
      is_primary_key:
        type: boolean
      is_unique:
        type: boolean
      name_en:
        type: string
      name_zh:
        type: string
      order_num:
        type: integer
    type: object
  models.TableSchemaRequest:
    properties:
      fields:
        items:
          $ref: '#/definitions/models.TableField'
        type: array
      interface_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      operation:
        description: create_table, alter_table, drop_table
        example: create_table
        type: string
      schema_name:
        example: user_basic_library
        type: string
      table_name:
        example: users
        type: string
    required:
    - interface_id
    - operation
    - schema_name
    - table_name
    type: object
...
  models.ThematicInterface:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      interface_config:
        $ref: '#/definitions/models.JSONB'
      is_table_created:
        type: boolean
      library_id:
        type: string
      name_en:
        type: string
      name_zh:
        type: string
      parse_config:
        $ref: '#/definitions/models.JSONB'
      status:
        type: string
      table_fields_config:
        $ref: '#/definitions/models.JSONB'
      thematic_library:
        allOf:
        - $ref: '#/definitions/models.ThematicLibrary'
        description: 关联关系
      type:
        description: realtime, batch
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.ThematicLibrary:
    properties:
      access_level:
        description: public, internal, private
        type: string
      authorized_roles:
        items:
          $ref: '#/definitions/models.JSONB'
        type: array
      authorized_users:
        items:
          $ref: '#/definitions/models.JSONB'
        type: array
      category:
        description: business, technical, analysis, report
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      domain:
        description: user, order, product, finance, marketing
        type: string
      id:
        type: string
      name_en:
        type: string
      name_zh:
        type: string
      publish_status:
        description: draft, published, archived
        type: string
      retention_period:
        type: integer
      source_libraries:
        items:
          $ref: '#/definitions/models.JSONB'
        type: array
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.JSONB'
        type: array
      update_frequency:
        description: realtime, hourly, daily, weekly, monthly
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: string
    type: object
...
info:
  contact: {}
  description: 智慧园区数据底座后台服务，提供数据采集、处理、存储、治理和共享功能
  title: 数据底座服务 API
  version: "1.0"
paths:
...
  /tables/manage-schema:
    post:
      consumes:
      - application/json
      description: 通过PgMetaApi动态创建、修改、删除数据库表结构
      parameters:
      - description: 表结构操作请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TableSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 管理数据库表结构
      tags:
      - 数据基础库
  /thematic-interfaces:
    post:
      consumes:
      - application/json
      description: 创建新的主题接口
      parameters:
      - description: 主题接口信息
        in: body
        name: thematic_interface
        required: true
        schema:
          $ref: '#/definitions/models.ThematicInterface'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ThematicInterface'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建主题接口
      tags:
      - 主题接口
  /thematic-interfaces/{id}:
    delete:
      description: 软删除主题接口（更新状态为inactive）
      parameters:
      - description: 主题接口ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除主题接口
      tags:
      - 主题接口
    get:
      description: 根据ID获取主题接口详细信息
      parameters:
      - description: 主题接口ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ThematicInterface'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取主题接口详情
      tags:
      - 主题接口
    put:
      consumes:
      - application/json
      description: 更新主题接口信息
      parameters:
      - description: 主题接口ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: thematic_interface
        required: true
        schema:
          $ref: '#/definitions/models.ThematicInterface'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新主题接口
      tags:
      - 主题接口
  /thematic-libraries:
    post:
      consumes:
      - application/json
      description: 创建新的数据主题库
      parameters:
      - description: 数据主题库信息
        in: body
        name: library
        required: true
        schema:
          $ref: '#/definitions/models.ThematicLibrary'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ThematicLibrary'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建数据主题库
      tags:
      - 数据主题库
  /thematic-libraries/{id}:
    delete:
      description: 删除数据主题库
      parameters:
      - description: 数据主题库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除数据主题库
      tags:
      - 数据主题库
    get:
      description: 根据ID获取数据主题库详细信息
      parameters:
      - description: 数据主题库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ThematicLibrary'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取数据主题库详情
      tags:
      - 数据主题库
    put:
      consumes:
      - application/json
      description: 更新数据主题库信息
      parameters:
      - description: 数据主题库ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: library
        required: true
        schema:
          $ref: '#/definitions/models.ThematicLibrary'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新数据主题库
      tags:
      - 数据主题库
  /thematic-libraries/{id}/publish:
    post:
      description: 将主题库状态更新为已发布
      parameters:
      - description: 数据主题库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 发布数据主题库
      tags:
      - 数据主题库
swagger: "2.0"